refactor czi reader
