+++ napari_imsmicrolink/utils/image.py
@@ -86,3 +86,28 @@ def centered_transform(
     rot_mat[:2, 2] = translation[:2]
 
     return rot_mat
+
+
+def guess_rgb(shape: Tuple[int, ...]) -> bool:
+    """
+    Guess if the passed shape comes from rgb data.
+    If last dim is 3 or 4 assume the data is rgb, including rgba.
+
+    Parameters
+    ----------
+    shape : list of int
+        Shape of the data that should be checked.
+
+    Returns
+    -------
+    bool
+        If data is rgb or not.
+    """
+    ndim = len(shape)
+    last_dim = shape[-1]
+    if ndim > 2 and last_dim < 5:
+        rgb = True
+    else:
+        rgb = False
+
+    return rgb
